scripts/requirements_web.txt
# Web application requirements
flask>=2.0.0
plotly>=5.0.0
python-dotenv>=0.19.0
firebase-admin>=6.0.0
pandas>=1.3.0
numpy>=1.21.0
python-dateutil>=2.8.0
